Move code from map_widget and custom_layer_dialog to util/maplibre
