+++ scwx-qt/source/scwx/qt/util/maplibre.hpp
@@ -37,6 +37,18 @@ glm::vec2 LatLongToScreenCoordinate(const QMapLibre::Coordinate& coordinate);
 void SetMapStyleUrl(const std::shared_ptr<map::MapContext>& mapContext,
                     const std::string&                      url);
 
+/**
+ * @brief Find the first layer which should be drawn above the radar products
+ *
+ * @param [in] styleLayers The layers of the style
+ * @param [in] drawBelow A list of RE2 compatible regex's describing the layers
+ * to draw below
+ *
+ * @return The first layer to be drawn above the radar products
+ */
+std::string FindMapSymbologyLayer(const QStringList&              styleLayers,
+                                  const std::vector<std::string>& drawBelow);
+
 } // namespace maplibre
 } // namespace util
 } // namespace qt
